example/requirements.txt
-e ..
django-bootstrap3==11.0.0
django-bootstrap4==0.0.8
django-debug-toolbar==1.11
django-filter==2.2
tablib==0.13.0
